feat: store item logs

+++ src/server/routers/item.ts
@@ -13,7 +13,7 @@ export const item = router({
         description: z.string().optional(),
       })
     )
-    .mutation(async ({ input }) => {
+    .mutation(async ({ input, ctx }) => {
       const { id } = await prisma.item.create({
         data: {
           name: input.itemName,
@@ -23,14 +23,41 @@ export const item = router({
           description: input.description,
         },
       });
+
+      await prisma.campaignLog.create({
+        data: {
+          changeType: "ADD_ITEM",
+          changeDescription: input.itemName,
+          user: { connect: { id: ctx.userId } },
+          campaign: { connect: { id: input.campaignId } },
+        },
+      });
+
       return id;
     }),
   delete: privateProcedure
     .input(z.object({ campaignId: z.string(), itemId: z.string() }))
-    .mutation(async ({ input }) => {
+    .mutation(async ({ input, ctx }) => {
+      const item = await prisma.item.findUnique({
+        where: { id: input.itemId },
+      });
+
+      if (!item) return;
+
       await prisma.item.delete({
         where: { id: input.itemId, campaignId: input.campaignId },
       });
+
+      await prisma.campaignLog.create({
+        data: {
+          changeType: "REMOVE_ITEM",
+          changeDescription: item.name,
+          user: { connect: { id: ctx.userId } },
+          campaign: { connect: { id: input.campaignId } },
+        },
+      });
+
+      return item.id;
     }),
   update: privateProcedure
     .input(
@@ -43,8 +70,8 @@ export const item = router({
         notes: z.string().optional(),
       })
     )
-    .mutation(async ({ input }) => {
-      const { id } = await prisma.item.update({
+    .mutation(async ({ input, ctx }) => {
+      const { id, name, campaignId } = await prisma.item.update({
         data: {
           name: input.itemName,
           notes: input.notes,
@@ -54,6 +81,16 @@ export const item = router({
         },
         where: { id: input.itemId },
       });
+
+      await prisma.campaignLog.create({
+        data: {
+          changeType: "ALTER_ITEM",
+          changeDescription: name,
+          user: { connect: { id: ctx.userId } },
+          campaign: { connect: { id: campaignId } },
+        },
+      });
+
       return id;
     }),
 });
